requirements.txt
django==3.2.6
gunicorn==20.0.4
psycopg2-binary==2.9.1
python-dotenv==0.19.0